Adding cross-build compatibility
